Definition of database
Has problem with mapstruct and lombok, no mapping properties in mapper.

+++ src/main/java/mx/bank/ms/currency_exchange/adapters/web/resources/AmountAndCurrency.java
@@ -2,13 +2,15 @@ package mx.bank.ms.currency_exchange.adapters.web.resources;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 /**
  * AmountAndCurrency
  */
 public record AmountAndCurrency(
 
         @JsonProperty("amount")
-        Double amount,
+        BigDecimal amount,
 
         @JsonProperty("currency")
         String currency
