+++ src/main/java/mx/bank/ms/currency_exchange/adapters/web/resources/ControlCurrencyExchangeTransactionResponse.java
@@ -40,7 +40,7 @@ public record ControlCurrencyExchangeTransactionResponse(
         @JsonProperty("resultingAmount")
         @JsonInclude(JsonInclude.Include.NON_ABSENT)
         @JsonSetter(nulls = Nulls.FAIL)
-        Double resultingAmount,
+        BigDecimal resultingAmount,
 
         @JsonProperty("createdAt")
         String createdAt,
