+++ src/main/java/mx/bank/ms/currency_exchange/adapters/web/resources/ExecuteCurrencyExchangeTransactionResponse.java
@@ -41,7 +41,7 @@ public record ExecuteCurrencyExchangeTransactionResponse(
         @JsonProperty("resultingAmount")
         @JsonInclude(JsonInclude.Include.NON_ABSENT)
         @JsonSetter(nulls = Nulls.FAIL)
-        Double resultingAmount,
+        BigDecimal resultingAmount,
 
         @JsonProperty("createdAt")
         String createdAt,
